@@ -6,7 +6,7 @@ description: Demonstrates how to use the mdsflutter plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -22,11 +22,11 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.5
 
-  provider: ^3.0.0
+  provider: ^6.0.4
 
-  permission_handler: ^5.0.1
+  permission_handler: ^9.2.0
 
 dev_dependencies:
   flutter_test:
