@@ -3,19 +3,19 @@ import 'package:mdsflutter_example/DeviceConnectionStatus.dart';
 import 'package:mdsflutter_example/DeviceModel.dart';
 
 class Device {
-  String _address;
-  String _name;
-  String _serial;
+  String? _address;
+  String? _name;
+  String? _serial;
   DeviceConnectionStatus _connectionStatus = DeviceConnectionStatus.NOT_CONNECTED;
 
-  Device(String name, String address) {
+  Device(String? name, String? address) {
     _name = name;
     _address = address;
   }
 
-  String get name => _name != null ? _name : "";
-  String get address => _address != null ? _address : "";
-  String get serial => _serial != null ? _serial : "";
+  String? get name => _name != null ? _name : "";
+  String? get address => _address != null ? _address : "";
+  String? get serial => _serial != null ? _serial : "";
   DeviceConnectionStatus get connectionStatus => _connectionStatus;
 
   void onConnecting() => _connectionStatus = DeviceConnectionStatus.CONNECTING;
